hwt: Store executable name instead of full path in trace context
When looking up symbols for IP range filtering, 'hwt_find_sym' uses
'pmcstat_symbol_search_by_name', passing 'tc->image_name' as the
name of the traced exectuable.
This will silently fail as 'pmcstat_symbol_search_by_name' uses
compares the passed image name with 'image->pi_name', which is
not a path but only the name of the executable.

This commit uses 'basename' to extract the name of the traced
executable from the path provided by the user and stores it
in 'tc->image_name'.