hwt: Pass struct hwt_vm pointer in hwt_backend_read
'hwt_backend_read' currently passes a 'cpu_id' integer to the backend.
This greatly complicates things when trying to use 'hwt_backend_read'
with MODE_THREAD since we don't have an easy way of knowing which
CPU a given thread is running on, if it's running at all.

This commit changes 'hwt_backend_read' to pass a pointer to
the 'struct hwt_vm' associated with the device. This makes it
trivial to read offsets in MODE_THREAD since we can easily
look up backend-specific thread data.