x86: Add routines for querying XSAVE feature information
This patch adds several routines that track and expose information about
various XSAVE-related features. More specifically, it adds the ability to
check whether a given XFEATURE is supported and which XSAVE extensions
are supported. Furthermore, it adds several routines for calculating the
size and offsets within a save area given a XSAVE feature bitmap.