bhyve/pci_emul.c: Introduce notion of fixed and hotpluggable slots
A virtual PCI slot is considered empty if no devices were attached to
it using the `-s` commandline flag. Such slots are eligible for runtime
emulated device hotplugging. Conversely, virtual slots with preconfigured
devices are treated as 'fixed' and are left alone for the remainder
of the virtual machine's lifecycle.

This is a prerequisite for emulated PCI device hotplugging as it
allows us to choose appropriate virtual PCI slots when hotplugging
an emulated PCI device.